Customized the SHELL command of pcbasic, and recompiled noeof.exe
The windows command "shell copy file1+file2 dest" changed from the x86 to
x64 bits distributions of windows.
In x86 the command will create file1 if it does not exist.
In x64 the command does not wokr if file1 does not exist.
This customized SHELL of pcbasic will emulate the old behavior, from
python code.
recompiled for 64bits, for the newer windows versions.
Also the noeof.exe present in the repo was compiled for a 16
bits compatible OS, like the dual 16/32 win98. Now it has been
recompiled for 64bi